MPORTANT!!!! Memory Allocation - NEW!!! Comments and code structure
Memory Allocation
I change memory allocation method. I substitute malloc() with calloc(),
an advantages is that memory is initialised to zero.  Instead of
allocate a vector of uint8,  it is possible to allocate the struct in
use. The big advantage is that we DO NOT NEED to create FREE MEMORY
function for each struct.

Comments and code structure
New comments and code structure. Should be more clear.